Base class ExplicitFVScheme for code factorization
