+++ include/samurai/schemes/fv/cell_based/explicit_cell_based_scheme.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "../../explicit_scheme.hpp"
+#include "../explicit_FV_scheme.hpp"
 #include "cell_based_scheme__lin_hom.hpp"
 #include "cell_based_scheme__nonlin.hpp"
 
@@ -10,44 +10,29 @@ namespace samurai
      */
     template <class cfg, class bdry_cfg>
     class Explicit<CellBasedScheme<cfg, bdry_cfg>, std::enable_if_t<cfg::scheme_type == SchemeType::LinearHomogeneous>>
+        : public ExplicitFVScheme<CellBasedScheme<cfg, bdry_cfg>>
     {
-        using scheme_t       = CellBasedScheme<cfg, bdry_cfg>;
-        using input_field_t  = typename scheme_t::input_field_t;
-        using output_field_t = typename scheme_t::output_field_t;
+        using base_class = ExplicitFVScheme<CellBasedScheme<cfg, bdry_cfg>>;
+
+        using scheme_t       = typename base_class::scheme_t;
+        using input_field_t  = typename base_class::input_field_t;
+        using output_field_t = typename base_class::output_field_t;
+
+        using base_class::scheme;
 
         static constexpr std::size_t field_size        = input_field_t::size;
         static constexpr std::size_t output_field_size = cfg::output_field_size;
         static constexpr std::size_t stencil_size      = cfg::scheme_stencil_size;
         static constexpr std::size_t center_index      = cfg::center_index;
 
-      private:
-
-        const scheme_t* m_scheme = nullptr;
-
       public:
 
         explicit Explicit(const scheme_t& scheme)
-            : m_scheme(&scheme)
+            : base_class(scheme)
         {
         }
 
-        auto& scheme() const
-        {
-            return *m_scheme;
-        }
-
-        auto apply_to(input_field_t& input_field) const
-        {
-            output_field_t output_field(scheme().name() + "(" + input_field.name() + ")", input_field.mesh());
-            output_field.fill(0);
-
-            update_bc(input_field);
-            apply(output_field, input_field);
-
-            return output_field;
-        }
-
-        void apply(output_field_t& output_field, input_field_t& input_field) const
+        void apply(output_field_t& output_field, input_field_t& input_field) const override
         {
             scheme().for_each_stencil_and_coeffs(
                 input_field,
@@ -74,40 +59,26 @@ namespace samurai
      */
     template <class cfg, class bdry_cfg>
     class Explicit<CellBasedScheme<cfg, bdry_cfg>, std::enable_if_t<cfg::scheme_type == SchemeType::NonLinear>>
+        : public ExplicitFVScheme<CellBasedScheme<cfg, bdry_cfg>>
     {
-        using scheme_t                                 = CellBasedScheme<cfg, bdry_cfg>;
-        using input_field_t                            = typename scheme_t::input_field_t;
-        using output_field_t                           = typename scheme_t::output_field_t;
+        using base_class = ExplicitFVScheme<CellBasedScheme<cfg, bdry_cfg>>;
+
+        using scheme_t       = typename base_class::scheme_t;
+        using input_field_t  = typename base_class::input_field_t;
+        using output_field_t = typename base_class::output_field_t;
+
+        using base_class::scheme;
+
         static constexpr std::size_t output_field_size = cfg::output_field_size;
-
-      protected:
-
-        const scheme_t* m_scheme = nullptr;
 
       public:
 
         explicit Explicit(const scheme_t& scheme)
-            : m_scheme(&scheme)
+            : base_class(scheme)
         {
         }
 
-        auto& scheme() const
-        {
-            return *m_scheme;
-        }
-
-        auto apply_to(input_field_t& input_field) const
-        {
-            output_field_t output_field(scheme().name() + "(" + input_field.name() + ")", input_field.mesh());
-            output_field.fill(0);
-
-            update_bc(input_field);
-            apply(output_field, input_field);
-
-            return output_field;
-        }
-
-        void apply(output_field_t& output_field, input_field_t& input_field) const
+        void apply(output_field_t& output_field, input_field_t& input_field) const override
         {
             scheme().for_each_stencil_center(
                 input_field,
